% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareModels.R
\name{CompareModels}
\alias{CompareModels}
\title{Comparar modelos de regressao Multipla}
\usage{
CompareModels(List)
}
\arguments{
\item{List}{Lista contendo os resultados da funcao `fit_regression` ajustados para diferentes modelos.}
}
\value{
A funcao apresenta os resultados dos ajustes dos modelos de regressao.
}
\description{
Esta funcao compara modelos ajustados pela funcao `fit_regression`.
}
\examples{
#dados sem repeticao
data("DadosMED")
modelo=y~a*x^b/(c^b+x^b)
chute=list(a=36,b=2,c=30)
Ajuste1=fit_regression(Data =DadosMED,model = modelo,start = chute,design=1)
plot_regression(Ajuste1,plot=2)
plot_regression(Ajuste1,plot=7,xlabel = "Dias",ylabel = "Altura (cm)")

#Modelo logistico
Ajuste2=fit_regression(Data =DadosMED,model = 7,start = c(a=30,b=5,c=0.1),design=1)
plot_regression(Ajuste2,plot=7,xlabel = "Dias",ylabel = "Altura (cm)")

CompareModels(list(Ajuste1,Ajuste2))
}
\references{
SILVEIRA, F.G.; SILVA, F.F.; CARNEIRO, P.S.; MALHADO, C.H.M.;  MUNIZ, J.A.
Analise de agrupamento na selecao de modelos de regressao nao-lineares para curvas de
crescimento de ovinos cruzados. Ciencia Rural, v.41, p.692- 698, 2011.
}
\seealso{
\code{\link{aov}}, \code{\link{lm}}, \code{\link{nls}}
}
\author{
Alcinei Mistico Azevedo
}
