% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DadosDBC.R
\docType{data}
\name{DadosDBC}
\alias{DadosDBC}
\title{Dados de um exemplo em DBC.}
\format{
Data.frame
}
\usage{
data(DadosDBC)
}
\description{
Dados de um exemplo no delineamento em blocos casualizados.
}
\references{
Canal do youtube: <https://www.youtube.com/channel/UCDGyvLCJnv9RtTY1YMBMVNQ>
}
\author{
Alcinei Mistico Azevedo: \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
