% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DadosDIC.R
\docType{data}
\name{DadosDIC}
\alias{DadosDIC}
\title{Dados de um exemplo em DIC.}
\format{
Data.frame
}
\usage{
data(DadosDIC)
}
\description{
Dados de um exemplo no delineamento inteiramente casualizado.
}
\references{
Canal do youtube: <https://www.youtube.com/channel/UCDGyvLCJnv9RtTY1YMBMVNQ>
}
\author{
Alcinei Mistico Azevedo: \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
