% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DadosMED.R
\docType{data}
\name{DadosMED}
\alias{DadosMED}
\title{Dados de um exemplo sem repeticao.}
\format{
Data.frame
}
\usage{
data(DadosMED)
}
\description{
Dados de um exemplo sem repeticao.
}
\references{
Canal do youtube: <https://www.youtube.com/channel/UCDGyvLCJnv9RtTY1YMBMVNQ>
}
\author{
Alcinei Mistico Azevedo: \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
