% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_regression.R
\name{fit_regression}
\alias{fit_regression}
\title{Ajuste de modelos de regressao nao linear}
\usage{
fit_regression(Data,model,start,design=1,verbose=TRUE)
}
\arguments{
\item{Data}{:Matriz contendo o conjunto de dados. A configuracao desta
matriz depende do design selecionado:\cr
1 : Para dados sem repeticao deve haver apenas duas colunas, a primeira com
identificacao dos tratamentos quantitativos e a segunda da variavel resposta.\cr
2 : Para dados de experimento em DIC deve haver apenas tres colunas, a primeira
 com identificacao dos tratamentos quantitativos, a segunda com identificacao
 das repeticoes e a terceira com a variavel resposta.\cr
3 : Para dados com repeticao de experimento em delineamento em blocos
casualizados deve haver apenas tres colunas, a primeira com identificacao dos
tratamentos quantitativos, a segunda com identificacao dos blocos e a terceira
com a variavel resposta.\cr}

\item{model}{:Refere-se ao modelo de regressao nao linear que deseja-se ajustar.
Pode ser uma formula ou um numero indicando um dos modelos abaixo:\cr
1: y ~ a/(1+b*exp(b-c*x))  -> Exponencial1\cr
2: y ~ a*exp(b*x) -> Exponencial2\cr
3: y~ a/(1+d*exp(b*c-x))^(1/d) -> Schnute\cr
4: y ~ a*(1-exp(c*b-c*x)) -> Michierlich\cr
5: y~a/(1+exp(b-c*x))^(1/d)  -> Richards\cr
6: y ~ a*exp(-exp(b-c*x)) -> Gompertz\cr
7: y ~ a/(1+exp(b-c*x))  -> Logistico\cr
8: y~a-b*exp(-c*x)  -> Meloun_I\cr
9: y~a-exp(-b-c*x) ->  Meloun_II\cr
10: y ~ a*(1-b*exp(-c*x))  ->  Brody\cr
11: y~a*(1-b*exp(-c*x))^3  ->  VonBertalanffy\cr
12: y ~ (a*x)/(x+b)   -> Michaelis_Menten\cr}

\item{start}{:valores iniciais (chutes) para o processo iterativo inseridos em um
vetor.}

\item{design}{:valor numerico que indica se:\cr
1 : para dados sem repeticao.\cr
2 : para dados com repeticao de experimento em delineamento
 inteiramente casualizado.\cr
3 : para dados com repeticao de experimento em delineamento em
blocos casualizados.\cr}

\item{verbose}{: Valor logico. Se TRUE (default) serao apresentados os
resultados da regressao.}
}
\value{
A funcao apresenta os resultados dos ajustes dos modelos de regressao.
}
\description{
Esta funcao realiza o ajuste de modelos de regressao nao linear
considerando dados sem repeticao ou com delineamento estatistico.
}
\examples{
#dados sem repeticao
data("DadosMED")
modelo=y~a*x^b/(c^b+x^b)
chute=list(a=36,b=2,c=30)
Ajuste1=fit_regression(Data =DadosMED,model = modelo,start = chute,design=1)
plot_regression(Ajuste1,plot=2)
plot_regression(Ajuste1,plot=7,xlabel = "Dias",ylabel = "Altura (cm)")

#Modelo logistico
Ajuste2=fit_regression(Data =DadosMED,model = 7,start = c(a=30,b=5,c=0.1),design=1)
plot_regression(Ajuste2,plot=7,xlabel = "Dias",ylabel = "Altura (cm)")


#dados em DIC
data(DadosDIC)
modelo=y~a*x^b/(c^b+x^b)
chute=list(a=36,b=2,c=30)
Ajuste3=fit_regression(Data =DadosDIC,model = modelo,start = chute,design=2)
plot_regression(Ajuste3,plot=7,xlabel = "Doses",ylabel = "Altura (cm)")

#
#dados em DBC
data(DadosDBC)
modelo=y~a*x^b/(c^b+x^b)
chute=list(a=36,b=2,c=30)
Ajuste3=fit_regression(Data =DadosDBC,model = modelo,start = chute,design=3)
plot_regression(Ajuste3,plot=7,xlabel = "Doses",ylabel = "PROD")

}
\references{
SILVEIRA, F.G.; SILVA, F.F.; CARNEIRO, P.S.; MALHADO, C.H.M.;  MUNIZ, J.A.
Analise de agrupamento na selecao de modelos de regressao nao-lineares para curvas de
crescimento de ovinos cruzados. Ciencia Rural, v.41, p.692- 698, 2011.
}
\seealso{
\code{\link{aov}}, \code{\link{lm}}, \code{\link{nls}}
}
\author{
Alcinei Mistico Azevedo
}
