% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonlinearRegression-package.R
\docType{package}
\name{NonlinearRegression package}
\alias{NonlinearRegression package}
\alias{NonlinearRegression}
\title{Pacote Para Analise De Experimentos Com regressão nao linear}
\description{
Este pacote realiza a analise de experimentos
com regressao nao linear considerando delineamento estatistico.
}
\references{
Canal do youtube: <https://www.youtube.com/channel/UCDGyvLCJnv9RtTY1YMBMVNQ>
 https://www.youtube.com/playlist?list=PLvth1ZcREyK72M3lFl7kBaHiVh5W53mlR

 HAIR, J.F. Multivariate Data Analysis.  (2016) 6ed. Pearson Prentice HalL.
  (ISBN 13:978 0138132637)
}
\author{
Alcinei Mistico Azevedo: <alcineimistico@hotmail.com>
}
