% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regression.R
\name{plot_regression}
\alias{plot_regression}
\title{Criacao de grafico de regressao nao linear}
\usage{
plot_regression(Adj,xlabel="x", ylabel="y", colPoints="blue",colLine="red",plot=7)
}
\arguments{
\item{Adj}{:Objeto criado pela funcao fit_regression.}

\item{xlabel}{:Titulo no eixo x.}

\item{ylabel}{:Titulo no eixo y.}

\item{colPoints}{:cor dos pontos de dispersao.}

\item{colLine}{: cor da linha.}

\item{plot}{: Valor numerico variando entre 1 e 9 indicando o visual do grafico.}
}
\value{
A funcao apresenta um grafico do modelo ajustado.
}
\description{
Esta funcao cria um grafico do ajuste do modelos de regressao nao linear.
}
\examples{
#dados sem repeticao
data("DadosMED")
modelo=y~a*x^b/(c^b+x^b)
chute=list(a=36,b=2,c=30)
Ajuste1=fit_regression(Data =DadosMED,model = modelo,start = chute,design=1)
plot_regression(Ajuste1,plot=2)
plot_regression(Ajuste1,plot=7,xlabel = "Dias",ylabel = "Altura (cm)")

#Modelo logistico
Ajuste2=fit_regression(Data =DadosMED,model = 7,start = c(a=30,b=5,c=0.1),design=1)
plot_regression(Ajuste2,plot=7,xlabel = "Dias",ylabel = "Altura (cm)")


#dados em DIC
data(DadosDIC)
modelo=y~a*x^b/(c^b+x^b)
chute=list(a=36,b=2,c=30)
Ajuste3=fit_regression(Data =DadosDIC,model = modelo,start = chute,design=2)
plot_regression(Ajuste3,plot=7,xlabel = "Doses",ylabel = "Altura (cm)")

#
#dados em DBC
data(DadosDBC)
modelo=y~a*x^b/(c^b+x^b)
chute=list(a=36,b=2,c=30)
Ajuste3=fit_regression(Data =DadosDBC,model = modelo,start = chute,design=3)
plot_regression(Ajuste3,plot=7,xlabel = "Doses",ylabel = "PROD")
}
\references{
SILVEIRA, F.G.; SILVA, F.F.; CARNEIRO, P.S.; MALHADO, C.H.M.;  MUNIZ, J.A.
Analise de agrupamento na selecao de modelos de regressao nao-lineares para curvas de
crescimento de ovinos cruzados. Ciencia Rural, v.41, p.692- 698, 2011.
}
\seealso{
\code{\link{aov}}, \code{\link{lm}}, \code{\link{nls}}
}
\author{
Alcinei Mistico Azevedo
}
